Extract test classes definition
