+++ spec/support/faulty_loader.rb
@@ -1,0 +1,5 @@
+class FaultyLoader < BaseLoader
+  def perform
+    raise 'test error'
+  end
+end
