+++ spec/support/faulty_normalizer.rb
@@ -1,0 +1,5 @@
+class FaultyNormalizer < BaseNormalizer
+  def call
+    raise
+  end
+end
