+++ spec/support/faulty_processor.rb
@@ -1,0 +1,5 @@
+class FaultyProcessor < BaseProcessor
+  def entities
+    raise 'test error'
+  end
+end
