+++ spec/support/test_processor.rb
@@ -1,0 +1,8 @@
+class TestProcessor < BaseProcessor
+  def entities
+    [
+      Entity.new(uid: '1', content: '', feed: Feed.new),
+      Entity.new(uid: '2', content: '', feed: Feed.new)
+    ]
+  end
+end
